pyproject.toml
[project]
name = "novel-outline-generator"
version = "0.1.0"
description = "Novel outline generator using LLM"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "zhipuai>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=. --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "tests/*",
    "outputs/*",
    "__pycache__/*",
    ".venv/*",
    "venv/*",
    "docs/*",
    "ui/dist/*",
]
